+++ staging/src/k8s.io/apiextensions-apiserver/pkg/startupcrd/configs/vmset_crd_definition.yaml
