+++ staging/src/k8s.io/apiextensions-apiserver/pkg/startupcrd/reader.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package crdinstall
+package startupcrd
 
 import (
 	"bytes"
